% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Create color palettes}
\usage{
get_color_palette(...)
}
\arguments{
\item{...}{Character names of colors}
}
\value{
A list of named vectors. Each vector is a color palette with each element containing the
color name and corresponding hex code
}
\description{
Create color palettes
}
\examples{
create_color_palette()
create_color_palette(...)
}
