% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_hex_color}
\alias{get_hex_color}
\title{Extract UCI colors as hex codes.}
\usage{
get_hex_color(...)
}
\arguments{
\item{...}{Character names of colors}
}
\value{
A named vector of colors with corresponding hex code
}
\description{
Default to return all colors from UCI color palette
}
\examples{
get_hex_color()
get_hex_color("uci-gold", "lime-green")
}
