% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_sample.R
\name{get_random_sample}
\alias{get_random_sample}
\title{Sample rows}
\usage{
get_random_sample(df, n, seed = 111)
}
\arguments{
\item{df}{A data frame}

\item{n}{Number of rows to randomly sample}

\item{seed}{Seed number. Use for reproducibility}
}
\value{
A data frame with n randomly selected rows.
}
\description{
Get a random sample of rows from a data frame
}
\examples{
get_random_sample(mtcars, n = 3)
}
