% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{interpolate_color}
\alias{interpolate_color}
\title{Interpolate colors}
\usage{
interpolate_color(palette = "main", reverse = FALSE, n, ...)
}
\arguments{
\item{palette}{Character name of the color palette}

\item{reverse}{Boolean indicating whether to reverse palette}

\item{n}{Integer indicating the number of colors to return after interpolating the given colors in the palette}

\item{...}{Additional arguments to colorRampPalette()}
}
\value{
A character vector of length n that contains colors interpolated from the palette
}
\description{
Interpolate colors
}
\examples{
interpolate_color(get_color_palette())
}
