% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_palette}
\alias{scale_fill_palette}
\title{Fill scale constructor for a color palette}
\usage{
scale_fill_palette(
  palette_name = "main",
  reverse = FALSE,
  discrete = TRUE,
  n = n,
  ...
)
}
\arguments{
\item{palette_name}{Character name of color palette s}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{n}{Integer indicating the number of colors to return after interpolating the given colors in the palette}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for a color palette
}
