% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti.R
\name{spaghetti}
\alias{spaghetti}
\title{Spaghetti plot}
\usage{
spaghetti(
  df,
  x_var,
  x_type,
  group_var,
  y_var,
  x_lab = "",
  y_lab = "",
  title = "",
  alpha = 1,
  color = "gray",
  size = 1,
  ...
)
}
\arguments{
\item{df}{A data frame}

\item{x_var}{X-variable. Specify as symbol}

\item{x_type}{Type of X-variable. Two options: continuous or factor}

\item{group_var}{Grouping variable. Specify as symbol.}

\item{y_var}{Y-variable. Specify as symbol}

\item{x_lab}{Label for x-axis}

\item{y_lab}{Label for x-axis}

\item{title}{Title of the plot}

\item{alpha}{Transparency of the background individual lines. 0 is invisible and 1 is normal.}

\item{...}{Additional arguments to aesthetic specification of ggplot}
}
\value{
ggplot plot object
}
\description{
Display a spaghetti plot using the ggplot2 engine.
}
\examples{
spaghetti(df = df, x_var = time, y_var = volume, group_var = participant_id)
}
