% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_berdcsc.R
\name{theme_berdcsc}
\alias{theme_berdcsc}
\title{Add berdcsc theme to ggplot2 object}
\usage{
theme_berdcsc(base_size = 20)
}
\arguments{
\item{base_size}{Maximum font size in the plot. Smaller font sizes are computed relative to the
base size}
}
\description{
This function allows you to add the berdcsc theme to your ggplot2 graphics
}
\examples{
line <- ggplot(line_df, aes(x = year, y = BP)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_berdcsc()
}
