% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{uci_colors}
\alias{uci_colors}
\title{Return UCI colors}
\usage{
uci_colors()
}
\value{
A named vector of UCI color names with corresponding hex codes
}
\description{
Return UCI colors
}
\examples{
uci_colors()
}
